del from local only, faster mig, itest--CI verifier
